+++ src/itdelatrisu/opsu/beatmap/BeatmapParser.java
@@ -547,7 +547,7 @@ public class BeatmapParser {
 
 							// calculate BPM
 							if (!timingPoint.isInherited()) {
-								int bpm = Math.round(60000 / timingPoint.getBeatLength());
+								int bpm = Math.round(60000 / (float) timingPoint.getBeatLength());
 								if (beatmap.bpmMin == 0) {
 									beatmap.bpmMin = beatmap.bpmMax = bpm;
 								} else if (bpm < beatmap.bpmMin) {
